requirements.txt
Django
djangorestframework
pillow
python-dotenv
django-cors-headers